test(formal): bound high-half fee rates
- Prove solver-costly high-half routing over uint8 amounts and fee rates
- Keep low-half proofs over uint16 inputs
- Preserve full allowed fee coverage in 10,000-run adversarial fuzzing

+++ test/formal/StaticsLaunchLiquidityHook.halmos.t.sol
@@ -100,11 +100,11 @@ contract StaticsLaunchLiquidityHookHalmosTest is SymTest, Test {
         _checkAfterSwap(amount, feeBps, true, true);
     }
 
-    function check_afterSwapRoutesExactUnspecifiedFeeExactInputOneForZero(uint8 amount, uint16 feeBps) public {
+    function check_afterSwapRoutesExactUnspecifiedFeeExactInputOneForZero(uint8 amount, uint8 feeBps) public {
         _checkAfterSwap(amount, feeBps, false, true);
     }
 
-    function check_afterSwapRoutesExactUnspecifiedFeeExactOutputZeroForOne(uint8 amount, uint16 feeBps) public {
+    function check_afterSwapRoutesExactUnspecifiedFeeExactOutputZeroForOne(uint8 amount, uint8 feeBps) public {
         _checkAfterSwap(amount, feeBps, true, false);
     }
